requirement.txt
numpy==1.26.4
pandas==2.2.3
matplotlib==3.10.3
seaborn==0.13.2
scikit-learn==1.6.1
tensorflow==2.10.1
streamlit==1.22.0
Pillow==11.2.1
streamlit==1.22.0